requirements.txt
opencv-python
pillow
pyyaml